refactor: polish agent package
